requirements.txt
django
django-debug-toolbar
django-extensions
ipython[notebook]
django-bootstrap3
pilkit
django -imagekit